+++ tools/cli/commands/deleteStack.js
@@ -1,8 +1,11 @@
-import fs from 'fs';
-import shell from 'shelljs';
-import { moveToModules } from '../helpers/util';
+import { moveToModules, handleDeleteDirectory, getPathsDirectory } from '../helpers/util';
 import { BASE_PATH } from '../config';
 
+/**
+ * Delete unused stack of technologies
+ *
+ * @param stackList - List unused stack of technologies
+ */
 const deleteStack = stackList => {
   const route = moveToModules('modules');
   const dirsList = getPathsDirectory(route);
@@ -12,25 +15,6 @@ const deleteStack = stackList => {
       handleDeleteDirectory(`${dir}/${stack === 'server' ? 'server-ts' : stack}`);
     });
   });
-};
-
-const handleDeleteDirectory = path => {
-  try {
-    shell.rm('-rf', path);
-  } catch (e) {
-    console.log('The stack was not found');
-  }
-};
-
-const getPathsDirectory = route => {
-  const dirPathList = [];
-  const elements = fs.readdirSync(route);
-  elements.forEach(element => {
-    if (!fs.statSync(`${route}/${element}`).isFile()) {
-      dirPathList.push(`${route}/${element}`);
-    }
-  });
-  return dirPathList;
 };
 
 export default deleteStack;
